refactor: split gui/mod.rs into focused submodules (actions, menu, rendering, hot_reload, window)
Pure refactor — no behavior change. Reduces gui/mod.rs from ~4976 to ~1850 lines by
extracting:
- actions.rs (374 lines): key action dispatch, zoom, pane/tab close, focus navigation
- menu.rs (424 lines): menu bar, pane context menu, tab bar, playback controls
- rendering.rs (474 lines): egui options, theme sync, PTY responses, clipboard, window manipulation
- hot_reload.rs (93 lines): shader and background image propagation to secondary windows
- window.rs (~1694 lines): SecondaryWindowState, spawn, frame loop, split pane
