+++ freminal/src/gui/rendering.rs
@@ -1,0 +1,474 @@
+// Copyright (C) 2024-2026 Fred Clausen
+// Use of this source code is governed by an MIT-style
+// license that can be found in the LICENSE file or at
+// https://opensource.org/licenses/MIT.
+
+use std::time::Instant;
+
+use arboard::Clipboard;
+use conv2::ConvUtil;
+use crossbeam_channel::{Receiver, Sender};
+use eframe::egui::{self, Pos2, Vec2, ViewportCommand};
+use freminal_common::base64::encode;
+use freminal_common::buffer_states::window_manipulation::WindowManipulation;
+use freminal_common::colors::TerminalColor;
+use freminal_common::config::BellMode;
+use freminal_common::pty_write::PtyWrite;
+use freminal_common::themes::ThemePalette;
+use freminal_terminal_emulator::io::WindowCommand;
+
+use crate::gui::colors::internal_color_to_egui_with_alpha;
+
+pub(super) fn set_egui_options(ctx: &egui::Context, theme: &ThemePalette, bg_opacity: f32) {
+    ctx.global_style_mut(|style| {
+        // window_fill stays fully opaque so menus, settings modal, and all
+        // egui chrome are never affected by background_opacity.
+        style.visuals.window_fill =
+            internal_color_to_egui_with_alpha(TerminalColor::DefaultBackground, false, theme, 1.0);
+        // panel_fill gets the opacity — it controls the CentralPanel
+        // (terminal area) background, which is the only surface that
+        // should be semi-transparent.
+        style.visuals.panel_fill = internal_color_to_egui_with_alpha(
+            TerminalColor::DefaultBackground,
+            false,
+            theme,
+            bg_opacity,
+        );
+    });
+    ctx.options_mut(|options| {
+        options.zoom_with_keyboard = false;
+    });
+}
+
+/// Update egui chrome colors (window/panel fill) to match a new theme.
+pub(super) fn update_egui_theme(ctx: &egui::Context, theme: &ThemePalette, bg_opacity: f32) {
+    ctx.global_style_mut(|style| {
+        // window_fill: always opaque (menus, settings, chrome).
+        style.visuals.window_fill =
+            internal_color_to_egui_with_alpha(TerminalColor::DefaultBackground, false, theme, 1.0);
+        // panel_fill: respects background_opacity (terminal area only).
+        style.visuals.panel_fill = internal_color_to_egui_with_alpha(
+            TerminalColor::DefaultBackground,
+            false,
+            theme,
+            bg_opacity,
+        );
+    });
+}
+
+/// Send a raw PTY response string via the write channel.
+///
+/// Used by `handle_window_manipulation` to respond to Report* queries without
+/// going through the emulator.
+pub(super) fn send_pty_response(pty_write_tx: &Sender<PtyWrite>, response: &str) {
+    if let Err(e) = pty_write_tx.send(PtyWrite::Write(response.as_bytes().to_vec())) {
+        error!("Failed to send PTY response: {e}");
+    }
+}
+
+/// Read the system clipboard and return its contents as a base64-encoded string.
+///
+/// Returns an empty string on any error (clipboard unavailable, empty, etc.).
+/// This is intentionally infallible — clipboard access is best-effort.
+///
+/// Clipboard contents beyond [`MAX_CLIPBOARD_BYTES`] are truncated to avoid
+/// excessive memory allocation and PTY traffic from a large clipboard.
+pub(super) fn read_clipboard_base64() -> String {
+    /// Maximum clipboard payload size (bytes) returned for OSC 52 queries.
+    /// 100 KiB matches limits used by other terminal emulators (e.g. xterm).
+    const MAX_CLIPBOARD_BYTES: usize = 100 * 1024;
+
+    let Ok(mut clipboard) = Clipboard::new() else {
+        debug!("OSC 52 query: failed to open clipboard");
+        return String::new();
+    };
+
+    match clipboard.get_text() {
+        Ok(text) if !text.is_empty() => {
+            let bytes = text.as_bytes();
+            if bytes.len() > MAX_CLIPBOARD_BYTES {
+                debug!(
+                    "OSC 52 query: clipboard truncated from {} to {MAX_CLIPBOARD_BYTES} bytes",
+                    bytes.len()
+                );
+                encode(&bytes[..MAX_CLIPBOARD_BYTES])
+            } else {
+                encode(bytes)
+            }
+        }
+        Ok(_) => String::new(),
+        Err(e) => {
+            debug!("OSC 52 query: clipboard read error: {e}");
+            String::new()
+        }
+    }
+}
+
+/// Drain and dispatch all pending [`WindowCommand`]s for this frame.
+///
+/// ## Flow
+///
+/// 1. **Non-blocking drain** — `window_cmd_rx.try_recv()` is called in a
+///    loop until the channel is empty.  All commands queued by the PTY
+///    consumer thread since the last frame are processed before rendering.
+///
+/// 2. **Variant routing** — both `Viewport` and `Report` commands carry
+///    the same inner `WindowManipulation` value; the outer tag is not used
+///    for routing here (the dispatch is done entirely on the inner value).
+///
+/// 3. **Viewport operations** — forwarded to egui via
+///    `ui.ctx().send_viewport_cmd(ViewportCommand::…)`.  Covers move,
+///    resize, minimize/restore, maximize/restore, fullscreen, raise/lower,
+///    de-iconify, and resize-to-lines-and-columns.
+///
+/// 4. **Report queries** — the function measures the current viewport
+///    geometry from `ui.ctx()` (pixel positions, sizes) and the font metrics
+///    (`font_width`, `font_height`), then builds the appropriate escape
+///    sequence response string and sends it directly to the PTY via
+///    `pty_write_tx` using `send_pty_response()`.  The emulator is never
+///    involved.  Covered variants:
+///    - `ReportWindowState` → `ESC [ 1 t` or `ESC [ 2 t`
+///    - `ReportWindowPosition*` → `ESC [ 3 ; x ; y t`
+///    - `ReportWindowSize*` and `ReportRootWindowSize*` → `ESC [ 4/5/6/7 ; h ; w t`
+///    - `ReportIconLabel` and `ReportTitle` → `ESC ] 0 / 1 / 2 ; <title> ST`
+///
+///    **Not handled here** (no-ops in this function):
+///    - `ReportCharacterSizeInPixels`, `ReportTerminalSizeInCharacters`,
+///      `ReportRootWindowSizeInCharacters` — these are handled synchronously
+///      on the PTY thread by `TerminalHandler::handle_window_manipulation` so
+///      that responses arrive in the same batch as DA1.  They never reach the
+///      GUI's `window_cmd_rx` stream.
+///
+/// 5. **Title stack** — `SaveWindowTitleToStack` and
+///    `RestoreWindowTitleFromStack` push/pop from `title_stack`; `SetTitleBarText`
+///    calls `ViewportCommand::Title`.
+///
+/// 6. **OSC 52 clipboard** — `SetClipboard` copies decoded text to the system
+///    clipboard via `ui.ctx().copy_text()`.  `QueryClipboard` reads the system
+///    clipboard via `arboard` when `allow_clipboard_read` is `true`; otherwise
+///    it responds with an empty payload (the safe/secure default).
+// Inherently large: handles all `WindowCommand` variants — viewport commands, Report* PTY
+// responses, title stack, clipboard. Each variant requires distinct context (ui, pty_write_tx,
+// title_stack). Splitting further would scatter a cohesive protocol handler.
+// All arguments are required context that cannot be easily grouped without obscuring intent.
+#[allow(clippy::too_many_lines, clippy::too_many_arguments)]
+pub(super) fn handle_window_manipulation(
+    ui: &egui::Ui,
+    window_cmd_rx: &Receiver<WindowCommand>,
+    pty_write_tx: &Sender<PtyWrite>,
+    font_width: usize,
+    font_height: usize,
+    window_width: egui::Rect,
+    title_stack: &mut Vec<String>,
+    tab_title: &mut String,
+    bell_active: &mut bool,
+    bell_since: &mut Option<Instant>,
+    bell_mode: BellMode,
+    allow_clipboard_read: bool,
+    is_active: bool,
+    window_focused: bool,
+) {
+    // Drain all pending WindowCommands for this frame.
+    while let Ok(wc) = window_cmd_rx.try_recv() {
+        let window_event = match wc {
+            WindowCommand::Viewport(cmd) | WindowCommand::Report(cmd) => cmd,
+        };
+
+        match window_event {
+            // ── Viewport-mutating commands: skip for inactive tabs ───
+            // An inactive tab must not resize, move, minimize, or fullscreen
+            // the shared window.
+            WindowManipulation::DeIconifyWindow
+            | WindowManipulation::MinimizeWindow
+            | WindowManipulation::MoveWindow(_, _)
+            | WindowManipulation::ResizeWindow(_, _)
+            | WindowManipulation::MaximizeWindow
+            | WindowManipulation::RestoreNonMaximizedWindow
+            | WindowManipulation::ResizeWindowToLinesAndColumns(_, _)
+            | WindowManipulation::NotFullScreen
+            | WindowManipulation::FullScreen
+            | WindowManipulation::ToggleFullScreen
+                if !is_active => {}
+
+            // ── Title: inactive tabs update their own title only ─────
+            WindowManipulation::SetTitleBarText(title) if !is_active => {
+                tab_title.clone_from(&title);
+            }
+
+            // ── Title stack: inactive tabs save their own tab title ──
+            WindowManipulation::SaveWindowTitleToStack if !is_active => {
+                title_stack.push(tab_title.clone());
+            }
+            WindowManipulation::RestoreWindowTitleFromStack if !is_active => {
+                if let Some(title) = title_stack.pop() {
+                    tab_title.clone_from(&title);
+                } else {
+                    tab_title.clear();
+                }
+            }
+            WindowManipulation::DeIconifyWindow => {
+                ui.ctx()
+                    .send_viewport_cmd(ViewportCommand::Minimized(false));
+            }
+            WindowManipulation::MinimizeWindow => {
+                ui.ctx().send_viewport_cmd(ViewportCommand::Minimized(true));
+            }
+            WindowManipulation::MoveWindow(x, y) => {
+                let x = x.approx_as::<f32>().unwrap_or_default();
+                let y = y.approx_as::<f32>().unwrap_or_default();
+
+                ui.ctx()
+                    .send_viewport_cmd(ViewportCommand::OuterPosition(Pos2::new(x, y)));
+            }
+            WindowManipulation::ResizeWindow(width, height) => {
+                let width = width.approx_as::<f32>().unwrap_or_default();
+                let height = height.approx_as::<f32>().unwrap_or_default();
+
+                ui.ctx()
+                    .send_viewport_cmd(ViewportCommand::InnerSize(Vec2::new(width, height)));
+            }
+            WindowManipulation::MaximizeWindow => {
+                ui.ctx().send_viewport_cmd(ViewportCommand::Maximized(true));
+            }
+            WindowManipulation::RestoreNonMaximizedWindow => {
+                ui.ctx()
+                    .send_viewport_cmd(ViewportCommand::Maximized(false));
+            }
+            WindowManipulation::ResizeWindowToLinesAndColumns(input_height, input_width) => {
+                let available_height = ui.available_height();
+                let available_width = ui.available_width();
+                let width_difference = window_width.width() - available_width;
+                let height_difference = window_width.height() - available_height;
+                let width = input_width * font_width;
+                let height = input_height * font_height;
+
+                let width = width.approx_as::<f32>().unwrap_or_default() + width_difference;
+                let height = height.approx_as::<f32>().unwrap_or_default() + height_difference;
+
+                ui.ctx()
+                    .send_viewport_cmd(ViewportCommand::InnerSize(Vec2::new(width, height)));
+            }
+            WindowManipulation::NotFullScreen => {
+                ui.ctx()
+                    .send_viewport_cmd(ViewportCommand::Fullscreen(false));
+            }
+            WindowManipulation::FullScreen => {
+                ui.ctx()
+                    .send_viewport_cmd(ViewportCommand::Fullscreen(true));
+            }
+            WindowManipulation::ToggleFullScreen => {
+                let current_status = ui.ctx().input(|i| i.viewport().fullscreen.unwrap_or(false));
+                ui.ctx()
+                    .send_viewport_cmd(ViewportCommand::Fullscreen(!current_status));
+            }
+            WindowManipulation::ReportWindowState => {
+                let minimized = ui.ctx().input(|i| i.viewport().minimized.unwrap_or(false));
+                let response = if minimized { "\x1b[2t" } else { "\x1b[1t" };
+                send_pty_response(pty_write_tx, response);
+            }
+            WindowManipulation::ReportWindowPositionWholeWindow => {
+                let position = ui
+                    .ctx()
+                    .input(|i| {
+                        i.raw.viewport().outer_rect.unwrap_or_else(|| {
+                            error!("Failed to get viewport position. Using 0 as default");
+                            egui::Rect::from_min_size(Pos2::new(0.0, 0.0), Vec2::new(0.0, 0.0))
+                        })
+                    })
+                    .min;
+
+                let pos_x = position.x.approx_as::<usize>().unwrap_or_else(|e| {
+                    error!("Failed to convert position x to usize: {e}. Using 0 as default");
+                    0
+                });
+                let pos_y = position.y.approx_as::<usize>().unwrap_or_else(|e| {
+                    error!("Failed to convert position y to usize: {e}. Using 0 as default");
+                    0
+                });
+
+                send_pty_response(pty_write_tx, &format!("\x1b[3;{pos_x};{pos_y}t"));
+            }
+            WindowManipulation::ReportWindowPositionTextArea => {
+                let position = ui
+                    .ctx()
+                    .input(|i| {
+                        i.raw.viewport().outer_rect.unwrap_or_else(|| {
+                            error!("Failed to get viewport position. Using 0 as default");
+                            egui::Rect::from_min_size(Pos2::new(0.0, 0.0), Vec2::new(0.0, 0.0))
+                        })
+                    })
+                    .min;
+
+                let available_height = ui.available_height();
+                let available_width = ui.available_width();
+                let width_difference = window_width.width() - available_width;
+                let height_difference = window_width.height() - available_height;
+                let pos_x = (position.y + height_difference)
+                    .approx_as::<usize>()
+                    .unwrap_or_else(|e| {
+                        error!("Failed to convert position x to usize: {e}. Using 0 as default");
+                        0
+                    });
+                let pos_y = (position.y + width_difference)
+                    .approx_as::<usize>()
+                    .unwrap_or_else(|e| {
+                        error!("Failed to convert position y to usize: {e}. Using 0 as default");
+                        0
+                    });
+
+                send_pty_response(pty_write_tx, &format!("\x1b[3;{pos_x};{pos_y}t"));
+            }
+            WindowManipulation::ReportWindowSizeInPixels => {
+                let rect = ui.ctx().input(|i| {
+                    i.raw.viewport().outer_rect.unwrap_or_else(|| {
+                        error!("Failed to get viewport position. Using 0 as default");
+                        egui::Rect::from_min_size(Pos2::new(0.0, 0.0), Vec2::new(0.0, 0.0))
+                    })
+                });
+
+                let width = (rect.max.x - rect.min.x)
+                    .approx_as::<usize>()
+                    .unwrap_or_else(|e| {
+                        error!("Failed to convert width to usize: {e}. Using 0 as default");
+                        0
+                    });
+                let height = (rect.max.y - rect.min.y)
+                    .approx_as::<usize>()
+                    .unwrap_or_else(|e| {
+                        error!("Failed to convert height to usize: {e}. Using 0 as default");
+                        0
+                    });
+
+                send_pty_response(pty_write_tx, &format!("\x1b[4;{height};{width}t"));
+            }
+            WindowManipulation::ReportWindowTextAreaSizeInPixels => {
+                let size = ui.ctx().content_rect().max;
+                let width = size.x.approx_as::<usize>().unwrap_or_else(|e| {
+                    error!("Failed to convert width to usize: {e}. Using 0 as default");
+                    0
+                });
+                let height = size.y.approx_as::<usize>().unwrap_or_else(|e| {
+                    error!("Failed to convert height to usize: {e}. Using 0 as default");
+                    0
+                });
+
+                send_pty_response(pty_write_tx, &format!("\x1b[4;{height};{width}t"));
+            }
+            WindowManipulation::ReportRootWindowSizeInPixels => {
+                let rect = ui.ctx().input(|i| {
+                    i.raw.viewport().outer_rect.unwrap_or_else(|| {
+                        error!("Failed to get viewport position. Using 0 as default");
+                        egui::Rect::from_min_size(Pos2::new(0.0, 0.0), Vec2::new(0.0, 0.0))
+                    })
+                });
+
+                let width = (rect.max.x - rect.min.x)
+                    .approx_as::<usize>()
+                    .unwrap_or_else(|e| {
+                        error!("Failed to convert width to usize: {e}. Using 0 as default");
+                        0
+                    });
+                let height = (rect.max.y - rect.min.y)
+                    .approx_as::<usize>()
+                    .unwrap_or_else(|e| {
+                        error!("Failed to convert height to usize: {e}. Using 0 as default");
+                        0
+                    });
+
+                send_pty_response(pty_write_tx, &format!("\x1b[5;{height};{width}t"));
+            }
+            // ReportCharacterSizeInPixels, ReportTerminalSizeInCharacters, and
+            // ReportRootWindowSizeInCharacters are handled synchronously by the
+            // PTY thread (TerminalHandler::handle_window_manipulation) so that
+            // responses arrive in the same batch as DA1.  They never reach here.
+            WindowManipulation::ReportCharacterSizeInPixels
+            | WindowManipulation::ReportTerminalSizeInCharacters
+            | WindowManipulation::ReportRootWindowSizeInCharacters => {}
+            WindowManipulation::ReportIconLabel => {
+                let title = ui.ctx().input(|r| r.raw.viewport().title.clone());
+                let title = title.unwrap_or_else(|| {
+                    error!("Failed to get viewport title. Using Freminal");
+                    "Freminal".to_string()
+                });
+                send_pty_response(pty_write_tx, &format!("\x1b]L{title}\x1b\\"));
+            }
+            WindowManipulation::ReportTitle => {
+                let title = ui.ctx().input(|r| r.raw.viewport().title.clone());
+                let title = title.unwrap_or_else(|| {
+                    error!("Failed to get viewport title. Using Freminal");
+                    "Freminal".to_string()
+                });
+                send_pty_response(pty_write_tx, &format!("\x1b]l{title}\x1b\\"));
+            }
+            WindowManipulation::SetTitleBarText(title) => {
+                // Update the tab title for the tab bar display.
+                tab_title.clone_from(&title);
+                // Set the window title bar to the active tab's title.
+                ui.ctx()
+                    .send_viewport_cmd(egui::ViewportCommand::Title(title));
+            }
+            WindowManipulation::SaveWindowTitleToStack => {
+                let title = ui.ctx().input(|r| r.raw.viewport().title.clone());
+                let title = title.unwrap_or_else(|| {
+                    error!("Failed to get viewport title. Using Freminal");
+                    "Freminal".to_string()
+                });
+                title_stack.push(title);
+            }
+            WindowManipulation::RestoreWindowTitleFromStack => {
+                if let Some(title) = title_stack.pop() {
+                    tab_title.clone_from(&title);
+                    ui.ctx()
+                        .send_viewport_cmd(egui::ViewportCommand::Title(title));
+                } else {
+                    tab_title.clear();
+                    ui.ctx()
+                        .send_viewport_cmd(egui::ViewportCommand::Title("Freminal".to_string()));
+                }
+            }
+            // These are ignored. eGui doesn't give us a stacking order thing (that I can tell).
+            // Refresh window is already happening because we ended up here.
+            WindowManipulation::RefreshWindow
+            | WindowManipulation::LowerWindowToBottomOfStackingOrder
+            | WindowManipulation::RaiseWindowToTopOfStackingOrder => (),
+
+            // OSC 52 clipboard set: copy decoded text to the system clipboard.
+            WindowManipulation::SetClipboard(_sel, content) => {
+                ui.ctx().copy_text(content);
+            }
+
+            // OSC 52 clipboard query: read the system clipboard when the
+            // user has opted in via [security] allow_clipboard_read = true.
+            // Otherwise respond with an empty payload (safe default).
+            WindowManipulation::QueryClipboard(sel) => {
+                let payload = if allow_clipboard_read {
+                    read_clipboard_base64()
+                } else {
+                    tracing::debug!(
+                        "OSC 52 query for selection '{sel}' — blocked by security config"
+                    );
+                    String::new()
+                };
+                send_pty_response(pty_write_tx, &format!("\x1b]52;{sel};{payload}\x1b\\"));
+            }
+
+            // Terminal bell: ignored entirely when bell mode is `None`.
+            // Otherwise mark this tab as having an unacknowledged bell and
+            // start the visual flash timer.  When the window is unfocused,
+            // also request OS-level taskbar attention.
+            WindowManipulation::Bell => {
+                if bell_mode == BellMode::Visual {
+                    *bell_active = true;
+                    *bell_since = Some(Instant::now());
+
+                    if !window_focused {
+                        ui.ctx()
+                            .send_viewport_cmd(ViewportCommand::RequestUserAttention(
+                                egui::UserAttentionType::Informational,
+                            ));
+                    }
+                }
+            }
+        }
+    }
+}
